Extract types to separate file
moves defualt layout function to root, as only used there.
