+++ src/index.tsx
@@ -5,4 +5,4 @@ export { default as Nav } from './Nav'
 export { default as NavListItem } from './NavListItem'
 export { default as Footer } from './Footer'
 export { default as presets } from './util'
-export * from './util'
+export * from './types'
