+++ src/util.tsx
@@ -1,103 +1,6 @@
-import { Breakpoint } from '@material-ui/core/styles/createBreakpoints'
-
-export type Variant = 'permanent' | 'persistent' | 'temporary'
-
-export type Position = 'static' | 'relative' | 'sticky' | 'fixed' | 'absolute'
-
-export type Orientation = 'left' | 'right'
-
-export type ScreenProps<P> = Partial<Record<Breakpoint, P>>
-
-export interface Layout {
-  clipped: boolean
-  collapsible: boolean
-  collapsedWidth: number
-  collapsed: boolean
-  footerShrink: boolean
-  navAnchor: Orientation
-  navVariant: Variant
-  navWidth: number
-  headerPosition: Position
-  open: boolean
-  squeezed: boolean
-  setCollapsed: (val: any) => any
-  setOpen: (val: any) => any
-}
-
-export interface LayoutConfig {
-  /**
-   * clipped, moves the header over the top of the navigation drawer
-   * unclipped makes navigation full height
-   * @default false
-   */
-  clipped: boolean | ScreenProps<boolean>
-  /**
-   * Can the navigation be collapsed to a smaller form
-   * @default true
-   */
-  collapsible: boolean | ScreenProps<boolean>
-  /**
-   * Width of the collapsed navigation
-   * @default 64
-   */
-  collapsedWidth: number | ScreenProps<number>
-  /**
-   * Footer to adjust the size to fit when nav expanded,
-   * set false to keep the same width and overflow the screen.
-   * @default true
-   */
-  footerShrink: boolean | ScreenProps<boolean>
-  /**
-   * Which side of the screen to show the nav panel
-   * @default left
-   */
-  navAnchor: Orientation | ScreenProps<Orientation>
-  /**
-   * Navigation variant:
-   * - permanent: stays all the time
-   * - persistent: remains open but can be hidden with button
-   * - temporary: hides on click away (and selection)
-   * @default permanent
-   */
-  navVariant: Variant | ScreenProps<Variant>
-  /**
-   * Width of the navigation drawer
-   * @default 256
-   */
-  navWidth: number | ScreenProps<number>
-  /**
-   * Position applied to the AppBar header
-   * one of 'static', 'relative', 'sticky', 'fixed', 'absolute'
-   * See https://developer.mozilla.org/en-US/docs/Learn/CSS/CSS_layout/Positioning
-   * @default relative
-   */
-  headerPosition: Position | ScreenProps<Position>
-  /**
-   * Both header and content adjust the size to fit when nav expanded,
-   * set false to keep the same width and overflow the screen.
-   * @default true
-   */
-  squeezed: boolean | ScreenProps<boolean>
-}
+import { Breakpoint, LayoutConfig, ScreenProps } from './types'
 
 const presets = {
-  defaultContext(): Layout {
-    return {
-      clipped: false,
-      collapsible: true,
-      collapsedWidth: 64,
-      collapsed: false,
-      footerShrink: true,
-      navAnchor: 'left',
-      navVariant: 'permanent',
-      open: true,
-      navWidth: 256,
-      headerPosition: 'relative',
-      squeezed: true,
-      setCollapsed: () => null,
-      setOpen: () => null
-    }
-  },
   createDefaultLayout(): LayoutConfig {
     return {
       clipped: false,
@@ -117,7 +20,7 @@ const presets = {
       squeezed: true
     }
   },
-  createFixedLayout: (config: LayoutConfig) => ({
+  createFixedLayout: (config: LayoutConfig): LayoutConfig => ({
     ...presets.createDefaultLayout(),
     navVariant: {
       xs: 'temporary',
@@ -132,7 +35,7 @@ const presets = {
     headerPosition: 'sticky',
     ...config
   }),
-  createContentBasedLayout: (config: LayoutConfig) => ({
+  createContentBasedLayout: (config: LayoutConfig): LayoutConfig => ({
     ...presets.createDefaultLayout(),
     navWidth: {
       sm: 200,
@@ -145,7 +48,7 @@ const presets = {
     collapsible: false,
     ...config
   }),
-  createCozyLayout: (config: LayoutConfig) => ({
+  createCozyLayout: (config: LayoutConfig): LayoutConfig => ({
     ...presets.createDefaultLayout(),
     navVariant: {
       xs: 'persistent',
@@ -163,7 +66,7 @@ const presets = {
     clipped: false,
     ...config
   }),
-  createMuiTreasuryLayout: (config: LayoutConfig) => ({
+  createMuiTreasuryLayout: (config: LayoutConfig): LayoutConfig => ({
     ...presets.createDefaultLayout(),
     navWidth: 200,
     navVariant: {
